firmware: fix dependency tracking for PBL firmware files
$(wildcard $(fwdir)/%) was added to allow for optional firmware, but in
return it broke dependency tracking, because it was evaluated before the
static pattern rule it's a part of.

Fix the dependency tracking by only evaluating it in a secondary
expansion. That this works as expected can be verified with:

  make -p | grep 'imx8mn-bl31.bin.gen.pbl.o:.*imx8mn-bl31.bin '

Which will return no result before this patch, but with it applied, it
will correctly show the dependency. Manual touching and rebuilding also
builds the firmware now as expected:

  AS [P]  firmware/imx8mm-bl31.bin.gen.pbl.o

Fixes: 5b171b7b7eda ("firmware: turn missing firmware into linker error")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230905142328.2935148-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>