ARM: i.MX8M: romapi: reserve scratch space
We have 32K of RAM at the end of the initial memory, where a
prebootloader can place data for future use. i.MX8M is currently the
only user and currently uses it to save the ROM API log, which is
located in SRAM and is inaccessible by the time barebox proper runs
in EL2. This area was so far not reserved, which can lead to clobbering
it when chainloading barebox. Avoid this by requesting the region.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230905142648.2936237-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>