scripts: Select XZ --arm64 BCJ filter for 64-bit arm
This change will select the arm64 BCJ filter for aarch64 instead of
the arm filter which was always selected for arm based architectures.

The selected filter is also tested, by looking if it is present in
the long help message of xz, this is because the arm64 BCJ filter is
a recent addition to xz.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230907152047.102747-2-jmaselbas@zdiv.net
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>