net: ksz9477: Add mdio bus slice dependency to i2c/spi device
On mdio busses the attached phys check for link status in a poller.
Add a slice dependency from the mdio bus to the I2C/SPI controller
so that we do not end up doing bus accesses while we are in a
I2C/SPI access on the same bus already.

Tested-by: Gerald Loacker <gerald.loacker@wolfvision.net>
Link: https://lore.barebox.org/20230908131117.1769957-3-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>