commands: test: support signed comparisons
So far signed integers in binary comparisons were accepted, but compared
unsigned resulting in [ "-1" -gt 0 ] returning true.

Rework this, so that:

  - If one side is signed, the comparison is signed
  - If both sides are unsigned, the comparison is unsigned

The latter is not required by POSIX, but it ensures that we don't break
users who so far did comparisons with integers > LONG_MAX without issue.

While at it, we start checking that the arguments in arithmetic
comparisons are indeed integers. This has the result that e.g.
[ "1x" -eq "1x" ] will no longer return success, but that aligns us with
usual shell behavior and protects the users from silent truncation when
a number > LONG_MAX is compared with a negative number.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230913115958.1858470-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>