ARM: ZynqMP: Add ZCU102 support
The ZCU102 is a potentially interesting platform, because there's Qemu
support for it. It's not very straight forward to use, because the
ZynqMP support in barebox and Linux relies heavily on firmware services,
which are lacking when naively using Qemu.

Still, let's add support now and worry about running it as part of the
test suite later.

Board support was not tested on actual hardware, but on Qemu with
changes on top of barebox to skip the firmware communication.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230913130150.2159921-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>