driver: add support for device aliases
Device names can get quite long, which makes them cumbersome to use
on the shell, e.g.  firmware:zynqmp-firmware:clock-controller.of.

In addition, the names are prone to change when the device tree nodes
are renamed.

One way we work around this is using aliases, but that is partially at
odds with upstream binding. This commit adds an alternative way of
allowing drivers and board code to set an alias that affects only
device_param_complete.

This provides an easy way of defining device names that should be stable
for use in shell scripts.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230913132456.2211919-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>