ARM: stm32mp: don't generate .img symlinks for .stm32 binaries
The .img extension for stm32mp1 images has been unfortunate. The format was
deprecated and finally removed in TF-A v2.8 in favor of the FIP format.

We won't remove support in barebox soon, but we shouldn't call these images
*.img. Since commit 3536e3904152 ("ARM: stm32mp: change stm32image extension to .stm32")
introduced with v2022.04.0, the .img files have been symlinks to regular files with
a stm32 extension and this commit now finishes the job and deletes the .img symlinks.

Users still using older TF-A versions should change their build system to use the same
barebox image file, but with .stm32 extension.

Users with new TF-A versions, should use the barebox-stm32mp-generic-bl33.img
as non-trusted firmware in the FIP in addition to a matching device tree generated
during the same barebox build.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230915130353.2217982-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>