ARM: i.MX: HABv4: point out i.MX6 specifics in docs and Kconfig help text
Signed USB images are a i.MX6 quirk as we need a point to DCD for
imx-usb-loader to work with, but the DCD pointer needs to be invalidated
for actual boot.

On i.MX8M*, RAM setup can't be done completely by DCD anymore, so not
having a DCD table is the norm and we don't need a signed USB image.

Thus mark CONFIG_HABV4_IMAGE_SIGNED_USB i.MX6 only and while at it, fix
hardcoded instances of i.MX6 headers.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230921075511.2098569-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>