startup: don't skip countdown if consoles were runtime enabled
Systems with verified boot will disable at least console input by
default, but when presented with an unlock token, they may elect to
enable the console. Currently, the user needs to be really quick with
ctrl+c to abort the boot on the active console, so let's make the system
have a normal count down if a console has been activated by board code.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230921092539.4142293-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>