test: self: have SELFTEST_ENABLE_ALL select all tests
CONFIG_SELFTEST_ENABLE_ALL is meant as a toggle to just enable all
tests that are applicable. imply doesn't do what's expected, because
it won't enable an option that's explicitly disabled. Thus switch
over to using select instead and while at it add the missing
select for SELFTEST_TEST_COMMAND.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231009115257.2371076-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>