uncompress: make use of error_fn in uncompress_buf_to_buf
uncompress_buf_to_buf already accepts an error_fn, but didn't make use
of it and instead relied on the default uncompress_err_stdout.

Fix this. The user-visible effect of this is that failure to decompress
FIT image contents is now printed with pr_err instead of the default
printf. No other upstream users of uncompress_buf_to_buf currently exist.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231009115310.2398292-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>