common: version: remove trailing new line from version_string
For determining the barebox version, the variables version_string,
release_string and buildsystem_version_string are exported for general
use. Only version_string is given a trailing new line though.

The trailing new line can be unexpected and is something that should
be added by code that requires it, so drop it from the variable
definition itself.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231010115906.3297847-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>