treewide: switch regmap.h include to linux/regmap.h
Now that there are no longer any users of regmap.h in headers, let's
switch all users in the source files to linux/regmap.h.

That way, the only users of regmap.h whether directly or indirectly will
be out-of-tree code, which will fail with an error if they are dependent
on the old semantics of regmap_bulk_read and regmap_bulk_write.

After a transitory period, we can then drop regmap.h.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231020071853.2826528-12-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>