ARM: stm32mp: disable stm32mp1_barebox_entry stack canary
stm32mp1_barebox_entry is given a stack canary when compiling PBL with
strong stack protector. The function can be called prior to relocation
though, so it needs to be marked as such.

This fixes boot with CONFIG_PBL_STACKPROTECTOR_STRONG=y on STM32MP1 with
generic BL33 entry point (for FIP).

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231101135811.4003573-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>