phy: stm32-usbphyc: fix unbalanced phy exit
On the STM32MP1, shutting down barebox can result in a

  ERROR: phy1: phy exit failed --> -22

message printed to the console. This is because the regulator is disabled
more often than it was enabled, because stm32_usbphyc_pll_disable
is called twice:

  - Once from stm32_usbphyc_remove calling stm32_usbphyc_phy_exit
  - Once from dwc2_remove calling phy_exit -> stm32_usbphyc_phy_exit

The code is taken from Linux and likely causes no issues there, because
devices are removed in reverse probe order, which barebox doesn't
enforce. Properly solving this would be a larger endeavour, so for now
just fix the stm32-usbphyc driver to ignore reference count dropping
below zero and only call __stm32_usbphyc_pll_disable once.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231109121012.1474536-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>