test: self: mmu: skip false positive test fail when no RAM at address 0
Testing zero_page_access only makes sense if we actually have RAM at
address 0 that we may read from. If it happens to I/O memory, the SoC
may hang or a data abort may be raised, which would make the test fail.

Avoid this by explicitly checking for that case and skipping the test
without running.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231109152946.2299755-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>