scripts: imx-image: don't pad the final binary for i.MX8M devices
If the padding was required the follwing error may appear:

| found i.MX8MP USB device [1fc9:0146]
| No dcd table in this ivt
| dl_command err=-1, last_trans=-1
| 4 in err=-4, last_trans=0  00 00 00 00

The error is triggered since the target request only the required size
of bytes and move on as soon as all bytes are received while, on the other
hand, host tools like imx-usb-loader and uuu try to send the complete
file size.

The alignment was introduced long time ago by commit 7cb4778e7f49
("imx-image: pad generated image to 4k") and may be required for HAB
boot on older SoCs like i.MX6/7. On these SoCs the CSF data is placed
behind the whole barebox image including the optional padding. On newer
SoCs like the i.MX8M* this isn't the case anymore. On these SoCs the CSF
is placed behind the barebox-pbl (aligned to 0x1000). So the final padding
isn't required anymore on these SoCs.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20231114133358.3746051-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>