console: define stub for console_get_first_active
console_get_first_active() is unconditionally called in startup.c to
check if the boot countdown can be skipped. The function is not defined
for CONFIG_CONSOLE_NONE however. Provide a stub in that case, so the
code can be compiled again.

This changes behavior: Configured countdown for systems without consoles
will be ignored and the system will boot directly. This is deemed
acceptable as it aligns behavior with systems with consoles that are
exclusively non-interactive and on the off-chance that a board requires
delays, it can always add a late init call delaying the boot.

Fixes: 66232c0ca70f ("startup: don't skip countdown if consoles were runtime enabled")
Reported-by: Steffen Hemer <S.Hemer@phytec.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20231120072122.2006805-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>